Extract some functionality from 'tell' to 'tell_fto' in fto-PageBot. Fix various misc. errors.

+++ page/invite.py
@@ -1,3 +1,6 @@
+import traceback
+import os.path
+
 import util
 
 INVITE_FILE = 'state/channel_invite.txt'
@@ -17,24 +20,23 @@ def h_invite(bot, id, target, channel, *args):
 
 @link('AUTOJOIN')
 def h_autojoin(bot):
-    try:
+    if os.path.exists(INVITE_FILE):
         with open(INVITE_FILE) as file:
             file_invited = map(str.strip, file.readlines())
-    except IOError:
-        file_invited = []
-    for chan in file_invited:
-        if chan.lower() in map(str.lower, bot.conf['channels']):
-            continue
-        bot.send_cmd('JOIN %s' % chan)
+        for chan in file_invited:
+            if chan.lower() in map(str.lower, bot.conf['channels']):
+                continue
+            bot.send_cmd('JOIN %s' % chan)
 
 @link('SELF_JOIN')
 def h_self_join(bot, chan):
     if chan.lower() not in invited:
         return
-    with open(INVITE_FILE) as file:
-        for line in file:
-            if chan.lower() == line.strip().lower():
-                return
+    if os.path.exists(INVITE_FILE):
+        with open(INVITE_FILE) as file:
+            for line in file:
+                if chan.lower() == line.strip().lower():
+                    return
     with open(INVITE_FILE, 'a') as file:
         file.write(chan + '\n')
     invited.remove(chan.lower())
@@ -43,12 +45,12 @@ def h_self_join(bot, chan):
 def h_self_part_kicked(bot, chan, *args):
     if chan.lower() in invited:
         invited.remove(chan.lower())
-    try:
+    if os.path.exists(INVITE_FILE):
         with open(INVITE_FILE) as file:
             file_invited = map(str.strip, file.readlines())
-    except IOError:
-        file_invited = []
-    new_file_invited = filter(lambda c: c.lower() != chan.lower(), file_invited)
-    if new_file_invited == file_invited: return
-    with open(INVITE_FILE, 'w') as file:
-        for c in new_file_invited: file.write(c + '\n')
+        new_file_invited = filter(
+            lambda c: c.lower() != chan.lower(), file_invited)
+        if new_file_invited == file_invited: return
+        with open(INVITE_FILE, 'w') as file:
+            for c in new_file_invited: file.write(c + '\n')
+
